Add some javadoc, mark a method as deprecated where we already have a simpler alternative and make the queries clearer and more explicit. Also remove some unnecessary conversion from Long to int to Integer. This is a fairly safe commit since it doesn't change any code paths or logic.
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@50943 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ common/common-services/src/main/java/ish/oncourse/services/course/ICourseService.java
@@ -62,7 +62,7 @@ public interface ICourseService {
 	 * Get courses count for current site. 
 	 * @return
 	 */
-	Integer getCoursesCount();
+	Long getCoursesCount();
 
 	/**
 	 * Get last modified date for course in current site.
